👷 Run unit tests as part of Jenkins pipeline
