+++ pkg/controller/unifiedpushserver/util.go
@@ -11,7 +11,6 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/discovery"
 )
 
 func labels(cr *pushv1alpha1.UnifiedPushServer, suffix string) map[string]string {
@@ -62,29 +61,4 @@ func updateContainerSpecImage(deployment *appsv1.Deployment, name string, image 
 			deployment.Spec.Template.Spec.Containers[idx].Image = image
 		}
 	}
-}
-
-// apiVersionChecker is just a container to make it easier to fake the
-// check function for tests
-type apiVersionChecker struct {
-	check func(apiGroupVersion string) (bool, error)
-}
-
-func getApiVersionChecker(dc discovery.DiscoveryInterface) *apiVersionChecker {
-	// Modified from https://github.com/operator-framework/operator-sdk/blob/947a464dbe968b8af147049e76e40f787ccb0847/pkg/k8sutil/k8sutil.go#L93
-	// The Operator Framework one checks a specific resource exists, but this function checks if an API version exists.
-	// Theoretically, there can be 2 resources in an API version, 1 exists and 1 not.
-	check := func(apiGroupVersion string) (bool, error) {
-		apiLists, err := dc.ServerResources()
-		if err != nil {
-			return false, err
-		}
-		for _, apiList := range apiLists {
-			if apiList.GroupVersion == apiGroupVersion {
-				return true, nil
-			}
-		}
-		return false, nil
-	}
-	return &apiVersionChecker{check: check}
 }
